Remove "304 not modified" Handling in http-Forwarder
This is a further change in scope of https://github.com/swisspush/gateleen/pull/180

Background:
Gateleen's Forwarder implicit compared ETags and decided to response with '304 not modified' and drain response body from upstream server. This reduces the amount of data to be transferred.
But it should be the Job of the upstream server to do correct Etag-handling and answer "304" and empty body accordingly. And in fact typical http servers like lighttpd or Swisspush's vertx-rest-storage already do this.
So there is in real life no need for Gateleed to do a second look on this Etag stuff.

The implementation in Gateleen in this case was problematic as it buffered the whole reponse body from the upstream server in-memory. This could lead to OOM-Exception. Instead of optimizing the code it was decided to simply remove the code and use the existing code path (which is already optimized in terms of memory consumption by using a Pump)