React on ungraceful closed client connections by closing server connection too
This is an addendum to avoid Exceptions in Formwarder when Gateleen's LocalHttpClient is used.

Background: Gateleen's LocalHttpClient is a shortcut 'withing Gateleen' to avoid localhost-Network-Traffic - used expecially in Gateleen's router. This 'shortcut' fakes a real HttpRequest/Response - but only a few methods of the appropriate Interfaces were implemented. We now need more methods to behave 'like it's real'