Keep RouteHook on rehooking when possible
When a RouteHook is re-registered, then the current implementation throws away the existing Route-objects - which finally contains the HttpClient.
This HttpClient is then closed forcefully which breaks ongoing http-Requests.

We change this by comparing the exising Route (resp. all the config properties for the Route) with the new config. If they are equal (with ecxeption of the now later expiry), then we keep the present Route (and HttpClient!) and just update the expiry.

Solves #250