Create correct HttpServerResponses also for LocalClient concerning handling of "Transfer-encoding: chunked" or "content-length" response headers.
Note that (at first glance) it seems not to be relevant for LocalClient as this is all 'in-memory'.
But we need a clear indication for the response in its headers if there is a body or not. Otherwise it's not possible (or very tedious) for the receiver of this response to detect "body or not".
Example: "304 response status for GET" or "200 ok for PUT" shall not (er even: must not) have a body - not even a "content-length" or a "transfer-encoding: chunked" header.