SDCISA-17956: Just a false-positive unluckily :(
I found ONE cause of these ConnectionClose issues. Unluckily, this one
is just a false-positive. It is just logged, but is not really an error
per se. The closing is even the desired/expected thing to happen in that
case. So the erorr is not really an error.

Therefore reduce distraction for this one by adjusting log level and
enhancing documentation. Then we have to hunt the remaining (still
unknown) cases.

Related:
    SDCISA-9578, SDCISA-9578, SDCISA-16515, SDCISA-17118, SDCISA-17118,
    SDCISA-17711, SDCISA-17711, SDCISA-17956, SDCISA-17986, SDCISA-20360,
    SDCISA-20740, SDCISA-20755, SDCISA-20763, SDCISA-20767, SDCISA-21066,
    ISAGDFPD-1553, gateleen PR 683, gateleen PR 686, gateleen PR 687,
    houston PR 816, gateleen PR 688.

(wip @ 1e8feed4bebe2f26a4273c04f072bc59c2757d7a)